+++ src/extension.ts
@@ -3,31 +3,30 @@
 
 import * as vscode from 'vscode';
 import { openTraceForActiveEditor, openTraceForFile } from './trace';
+import { Commands, TraceOpenResult } from './constants';
 
 export function activate(context: vscode.ExtensionContext): void {
 	console.log("activate() called");
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand(
-			"perfetto-trace.open-trace-active-editor",
-			() => {
-				openTraceForActiveEditor(context).then(
-					() => vscode.window.showInformationMessage("Successfully Opened Trace For Active Editor"),
-					(reason: string) => vscode.window.showErrorMessage(reason)
-				);
-			}
+			Commands.OpenTraceActiveEditor,
+			() => openTraceForActiveEditor(context).then(event => {
+				if (event !== TraceOpenResult.Success) {
+					vscode.window.showErrorMessage(`Failed to launch with reason: ${event}`);
+				}
+			})
 		)
 	);
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand(
-			"perfetto-trace.open-trace-file",
-			() => {
-				openTraceForFile(context).then(
-					() => vscode.window.showInformationMessage("Successfully Opened Trace For Chosen File"),
-					(reason: string) => vscode.window.showErrorMessage(reason)
-				);
-			}
+			Commands.OpenTraceFile,
+			() => openTraceForFile(context).then(event => {
+				if (event !== TraceOpenResult.Success) {
+					vscode.window.showErrorMessage(`Failed to launch with reason: ${event}`);
+				}
+			})
 		)
 	);
 }